breaking changes
Introducing CommitServiceV1 (not to be confused with CommitServiceV2)

The V1 service is an HTML scraper which can no longer be used. All the
design and tests are based around what it can do, with two main public
methods that are agnostic about how they actually get implemented.

The V2 service is a JSON API client which goes with the modern React
implementation of CommitsTo. It takes advantage of the unauthenticated,
unauthorized API to achieve the same things (without an HTML scraper.)

V2 service object is not yet written or tested