remove fancy shit
All of the process forking / scheduler model, etc things seem to just
get in the way. What we have here is a Ruby script that just runs.

IDK if it is efficient but I know it isn't threaded or fibrous or
whatever, if I need a different thread then I will run it in a separate
process and they can communicate with IPC, in the form of HTTP requests

(that's where it'll get complicated, because a process can't just do
what we're doing here AND listen for HTTP requests...)

Signed-off-by: Kingdon Barrett <kingdon@weave.works>